Refactor powerofnumrecursive class to optimize power calculation
The power method in the RecursiveReduced class has been modified to optimize the calculation of power. It now checks if the exponent (exp) is even or odd. If exp is even, it calculates the power of num raised to exp/2 and multiplies the result by itself. If exp is odd, it multiplies num by the result of power(num, exp - 1). This approach reduces the time complexity to O(log(n)).